Code Restructure, FireEye Functionality
